This is the title
And this is the body. It is very long. Will Github include it in its JSON? I don't know, but I hope so! Here's some random text: Though not required, it's a good idea to begin the commit message with a single short (less than 50 character) line summarizing the change, followed by a blank line and then a more thorough description. The text up to the first blank line in a commit message is treated as the commit title, and that title is used throughout git. For example, git-format-patch(1) turns a commit into email, and it uses the title on the Please enter the commit message for your changes. Lines starting with '#' will be ignored, and an empty message aborts the commit.